qute_core/qute_gateway_test/requirements.txt
aliyun-python-sdk-core==2.13.5
aliyun-python-sdk-core-v3==2.13.3
aliyun-python-sdk-kms==2.7.1
allure-python-commons==2.7.1
apipkg==1.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
configparser==3.7.4
crcmod==1.7
elasticsearch==7.0.2
enum34==1.1.6
execnet==1.7.0
idna==2.8
importlib-metadata==0.19
jmespath==0.9.4
lxml==4.4.1
more-itertools==7.2.0
namedlist==1.7
oss2==2.7.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
pycryptodome==3.8.2
pymongo==3.8.0
PyMySQL==0.7.11
pyparsing==2.4.2
pytest==5.0.1
pytest-allure-adaptor2==1.7.11
pytest-forked==1.0.2
pytest-html==1.21.1
pytest-metadata==1.8.0
pytest-xdist==1.29.0
redis==2.10.6
requests==2.21.0
six==1.12.0
urllib3==1.24.3
wcwidth==0.1.7
zipp==0.5.2
